Add files via upload
Can you see the logger file now? I was thinking of adding utility method in the UI file, but I wasn't sure if you like the Logger file, I needed to ask the Oracle (Chat) on how to do the LocalDateTime and DateTimeFormatter, it helped a little but I think I tweaked it a bit better, let me know your thoughts!